+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Fixed
 - Fixed an error that occurred when creating a width field on Craft 3.7.46 or later Craft 3 releases
+- Fixed a bug where Roll Your Own widgets were always displaying 'Roll Your Own' as the title, instead of the user-specified title
 
 ## 2.2.3 - 2022-05-24